Feat: MyApp으로 분리
