+++ app/src/main/java/inu/appcenter/intip_android/MyApp.kt
@@ -1,0 +1,32 @@
+package inu.appcenter.intip_android
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import inu.appcenter.intip_android.ui.navigate.AllDestination
+import inu.appcenter.intip_android.ui.web.WebViewScreen
+
+@Composable
+fun MyApp(modifier: Modifier = Modifier) {
+    val navController = rememberNavController()
+    NavHost(
+        navController = navController,
+        startDestination = AllDestination.Main.route,
+        modifier = modifier
+    ) {
+        composable(AllDestination.Main.route) {
+            WebViewScreen(url = "https://intip.inuappcenter.kr/app/home")
+        }
+        composable(AllDestination.Save.route) {
+            WebViewScreen(url = "https://intip.inuappcenter.kr/app/save")
+        }
+        composable(AllDestination.Write.route) {
+            WebViewScreen(url = "https://intip.inuappcenter.kr/app/write")
+        }
+        composable(AllDestination.MyPage.route) {
+            WebViewScreen(url = "https://intip.inuappcenter.kr/app/mypage")
+        }
+    }
+}
