+++ app/src/main/java/inu/appcenter/intip_android/ui/component/AppBottomBar.kt
@@ -1,4 +1,4 @@
-package inu.appcenter.intip_android.ui.navigate
+package inu.appcenter.intip_android.ui.component
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.background
@@ -30,6 +30,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import inu.appcenter.intip_android.R
 import inu.appcenter.intip_android.ui.modifier.shadowCustom
+import inu.appcenter.intip_android.ui.navigate.AllDestination
 import inu.appcenter.intip_android.ui.theme.Blue_9CAFE2
 import inu.appcenter.intip_android.ui.theme.Gray_D6D1D5
 import inu.appcenter.intip_android.ui.theme.INTIPTheme
